refactor: property issue detail
