+++ frontend/src/pages/project/project.tsx
@@ -83,17 +83,21 @@ export const ProjectPage = () => {
       <Pane paddingY={majorScale(2)}>
         <Table>
           <Table.Head>
-            <Table.TextHeaderCell>ID</Table.TextHeaderCell>
+            <Table.TextHeaderCell flexBasis={majorScale(8)} flexShrink={0} flexGrow={0}>
+              ID
+            </Table.TextHeaderCell>
             <Table.TextHeaderCell>Dự án</Table.TextHeaderCell>
             <Table.TextHeaderCell>Trạng thái</Table.TextHeaderCell>
             <Table.TextHeaderCell>Ngày bắt đầu</Table.TextHeaderCell>
             <Table.TextHeaderCell>Ngày kết thúc</Table.TextHeaderCell>
-            <Table.TextHeaderCell>Chỉnh sửa</Table.TextHeaderCell>
+            <Table.TextHeaderCell>Hành động</Table.TextHeaderCell>
           </Table.Head>
           <Table.Body>
             {projects.map((project) => (
               <Table.Row key={project.id}>
-                <Table.TextCell>{project.id}</Table.TextCell>
+                <Table.TextCell flexBasis={majorScale(8)} flexShrink={0} flexGrow={0}>
+                  {project.id}
+                </Table.TextCell>
                 <Table.TextCell>
                   <Pane display='flex' alignItems='center'>
                     <Avatar src={project.icon} marginRight={majorScale(1)} />
